Fix #7 - Move U2F protocol functions out of lib.rs into u2fprotocol.rs
